refactor: merge crates with a fluentbase version
