--- bin/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "rwasm-bin"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-rwasm-codegen = { path = "../codegen" }
-rwasm = { path = "../rwasm" }
-# we need runtime here only to obtain sys call indices
-fluentbase-runtime = { git = "https://github.com/fluentlabs-xyz/fluentbase", branch = "devel" }
-wat = "1.0.80"
-clap = { version = "4.4.11", features = ["derive"] }
-lazy_static = "1.4.0"
-log = "0.4.20"
-ctor = "0.2.6"
-env_logger = "0.11.0"
