--- bin/src/types.rs
@@ -1,5 +1,0 @@
-pub(crate) const RWASM_OUT_FILE_EXT: &str = ".rwasm";
-pub(crate) enum FileFormat {
-    Wat,
-    Wasm,
-}
