+++ e2e/Cargo.toml
@@ -4,7 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+rwasm-executor = { path = "../executor" }
 rwasm = { path = "../rwasm" }
+
 byteorder = "1.4.3"
 wat = "1.0.66"
 anyhow = "1.0.71"
@@ -13,4 +15,4 @@ serial_test = "3.0.0"
 
 [features]
 default = ["e2e"]
-e2e = ["rwasm/e2e"]
+e2e = ["rwasm-executor/e2e", "rwasm/e2e"]
