+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2023-11-05
+nightly-2025-01-27
